pyproject.toml
[project]
name = "chat-client-toy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.20",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "anthropic>=0.79.0",
    "sqlalchemy>=2.0.48",
    "tiktoken>=0.11.0",
    "pymupdf>=1.26.4",
    "PyPDF2>=3.0.1",
    "pyyaml>=6.0.2",
]
